Restructure object creation commands into a single command per object type.
About 98% of the code is shared between create/define commands, and the
distinction is lost on most users who have limited experience with
libvirt.

Unifying the `create`/`define` commands as a `new` command with an option
to specify whether objects should be ‘created’ or ‘defined’ makes
us a bit friendlier for new users, gets rid of some code duplication
(almost all of it was deduplicated already), and also allows easier
extension to support additional features like generating the object XML
config on-the-fly with our templating support.