Restructure imports to improve load times.
Many of our imports are only needed when specific functions are run. By
restructuring things so that those imports are made only when needed,
instead of always being made at the top of the file, load times for the
fvirt CLI tool can be improved significantly, especially when displaying
help text.